+++ official/legacy/image_classification/classifier_trainer_test.py
@@ -53,6 +53,7 @@ def distribution_strategy_combinations() -> Iterable[Tuple[Any, ...]]:
       model=[
           'efficientnet',
           'resnet',
+          'vgg',
       ],
       dataset=[
           'imagenet',
@@ -149,6 +150,7 @@ class ClassifierTest(tf.test.TestCase, parameterized.TestCase):
           model=[
               'efficientnet',
               'resnet',
+              'vgg',
           ],
           dataset='imagenet',
           dtype='float16',
@@ -193,6 +195,7 @@ class ClassifierTest(tf.test.TestCase, parameterized.TestCase):
           model=[
               'efficientnet',
               'resnet',
+              'vgg',
           ],
           dataset='imagenet',
           dtype='bfloat16',
