+++ official/legacy/image_classification/configs/examples/vgg16/gpu.yaml
