+++ official/legacy/image_classification/vgg16/__init__.py
